@@ -1,27 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Autenticacao" Version="0.0.2" />
     <PackageReference Include="AutoMapper" Version="9.0.0" />
-    <PackageReference Include="Conexoes" Version="0.2.9" />
+    <PackageReference Include="Conexoes" Version="0.3.10" />
+    <PackageReference Include="ConfiguracoesAPI" Version="0.0.4" />
     <PackageReference Include="Mensagens" Version="0.1.4" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
     <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.7.1" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.4" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="Retornos" Version="0.4.15" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Abstracoes\Abstracoes.csproj" />
-    <ProjectReference Include="..\Repositorios\Repositorios.csproj" />
-    <ProjectReference Include="..\Servicos\Servicos.csproj" />
+    <ProjectReference Include="..\Infraestrutura\Repositorio\Repositorio.csproj" />
+    <ProjectReference Include="..\Infraestrutura\Servico\Servico.csproj" />
   </ItemGroup>
 </Project>
