@@ -1,52 +1,50 @@
-using Abstracoes.Representacoes.Usuario.Grupo;
-using Abstracoes.Representacoes.Usuario.Pessoa;
-using Abstracoes.Representacoes.Usuario.Usuario;
+using Infraestrutura.Servico.Usuario.Entidade;
+using Infraestrutura.Servico.Usuario.Interface;
+using MandradePkgs.Autenticacao.Estrutura.Token;
 using MandradePkgs.Retornos;
 using Microsoft.AspNetCore.Mvc;
-using Servicos.Usuario.Interfaces;
 
 namespace Api.Controllers.Usuarios
 {
+    [Route("/usuarios")]
     [Produces("application/json")]
     [ApiController]
     public class UsuariosController : ControllerApi
     {
-        private IUsuarioSrv _servico { get; }
+        private IUsuarioSv _servico { get; }
 
-        public UsuariosController(IUsuarioSrv servico)
+        public UsuariosController(IUsuarioSv servico)
         {
             _servico = servico;
         }
 
+        [HttpPost]
+        [Route("autenticar")]
+        public RespostaApi<Token> Autenticar(string Usuario, string Senha, [FromServices] ConfiguracoesToken configsToken, [FromServices] AssinaturaToken assinatura) =>
+            RespostaPadrao(_servico.Autenticar(Usuario, Senha, configsToken, assinatura));
+
         [HttpGet]
-        [Route("/usuarios/{id}/pessoa")]
+        [Route("{id}/pessoa")]
         public RespostaApi<PessoaDto> PesquisarPessoaUsuario(int id) =>
             RespostaPadrao(_servico.ObterPessoaUsuario(id));
 
         [HttpGet]
-        [Route("/usuarios/{id}/grupo")]
+        [Route("{id}/grupo")]
         public RespostaApi<GrupoDto> PesquisarGrupoUsuario(int id) =>
             RespostaPadrao(_servico.ObterGrupoUsuario(id));
 
         [HttpPost]
-        [Route("/usuarios/")]
         public RespostaApi Cadastrar(UsuarioInclusaoDto usuario) =>
             RespostaPadrao(_servico.IncluirUsuario(usuario));
 
         [HttpPut]
-        [Route("/usuarios/{id}")]
+        [Route("{id}")]
         public RespostaApi AlterarAtividade(int id, bool ativo) =>
             RespostaPadrao(_servico.AtualizarAtividadeUsuario(id, ativo));
 
         [HttpPut]
-        [Route("/usuarios/{id}/alterar-senha")]
+        [Route("{id}/alterar-senha")]
         public RespostaApi AlterarSenha(int id, UsuarioAlteracaoSenhaDto alteracao) =>
             RespostaPadrao(_servico.AtualizarSenhaUsuario(id, alteracao));
-
-        [HttpDelete]
-        [Route("/usuarios/{id}")]
-        public RespostaApi Excluir(int id) =>
-            RespostaPadrao(_servico.ExcluirUsuario(id));
-
     }
 }
